Added some more solved problems

+++ C++ Programming Language/Task-1/sum_of_factors.cpp
@@ -61,9 +61,9 @@ int main (int argc, const char** argv) {
             }
       }
 
-      std::cout << f_sum (B, A, argc-1) << '\n';
+      std::cout << f_sum (B, A, static_cast<size_t> (argc-1)) << '\n';
 
-      for (size_t i = 0; i < argc-1; ++i) {
+      for (int i = 0; i < argc-1; ++i) {
             delete [] B[i];
       }
